added pagelaouts for graph
