+++ bitte.dev.json
@@ -1,5 +1,5 @@
 {
-  "url": "https://sixty-areas-pump.loca.lt",
-  "pluginId": "sixty-areas-pump.loca.lt",
+  "url": "https://wicked-symbols-shout.loca.lt",
+  "pluginId": "wicked-symbols-shout.loca.lt",
   "receivedId": ""
 }
